+++ src/main/java/de/samply/teiler/response/entity/RequestResponseEntity.java
@@ -1,7 +1,7 @@
 package de.samply.teiler.response.entity;
 
 public record RequestResponseEntity(
-    Long queryExecutionId
+    String responseUrl
 ) {
 
 }